GitLab still changes the path
Even if the path is explicitly sent with the data packet, GitLab still
modifies it for no reason.

I've removed the optional, "path" I added and replaced the dot with a
single underscore. GitLab has made changes that will appear in 10.3 to
allow the double underscore again. We may come back to this at that
time.